move code into src/; update scripts

+++ gwh_run48.sh
@@ -11,17 +11,12 @@ if [ ! -e $STATEDIR/gxrom-r ]; then
     )
 fi
 
-RAM="-ram ram"
+RAM=''
 if [ ! -e $STATEDIR/ram ]; then
-    RAM="-reset"
+    RAM=-reset
 fi
 
-if [ ! -e $STATEDIR/port1 ]; then
-    dd if=/dev/zero of=$STATEDIR/port1 bs=1k count=128
-fi
+[ ! -e $STATEDIR/port1 ] && dd if=/dev/zero of=$STATEDIR/port1 bs=1k count=128
+[ ! -e $STATEDIR/port2 ] && dd if=/dev/zero of=$STATEDIR/port2 bs=1k count=1024
 
-if [ ! -e $STATEDIR/port2 ]; then
-    dd if=/dev/zero of=$STATEDIR/port2 bs=1k count=1024
-fi
-
-./run_saturn -face hp48 -hw hp48 -stateDir $STATEDIR -rom gxrom-r "$RAM" -port1 port1 -port2 port2
+./run_saturn -face hp48 -hw hp48 -stateDir $STATEDIR -rom gxrom-r $RAM -port1 port1 -port2 port2
